Split session module
