+++ mqtt/.gitignore
@@ -1,1 +1,1 @@
-state.*
+state*.dat
