+++ mqtt/mqtt-broker/src/session/identifiers.rs
@@ -1,0 +1,236 @@
+use std::{cmp, fmt, mem};
+
+use serde::{
+    de::{SeqAccess, Visitor},
+    ser::SerializeTuple,
+    Deserialize, Deserializer, Serialize, Serializer,
+};
+
+use mqtt3::proto;
+
+use crate::Error;
+
+#[derive(Clone)]
+pub(crate) struct IdentifiersInUse(pub(crate) Box<[usize; PacketIdentifiers::SIZE]>);
+
+impl fmt::Debug for IdentifiersInUse {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("IdentifiersInUse").finish()
+    }
+}
+
+impl cmp::PartialEq for IdentifiersInUse {
+    fn eq(&self, other: &Self) -> bool {
+        self.0.iter().zip(other.0.iter()).all(|(a, b)| a.eq(b))
+    }
+}
+
+struct IdentifiersInUseVisitor;
+
+impl<'de> Visitor<'de> for IdentifiersInUseVisitor {
+    type Value = IdentifiersInUse;
+
+    fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(formatter, "an array of length {}", PacketIdentifiers::SIZE)
+    }
+
+    #[inline]
+    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
+    where
+        A: SeqAccess<'de>,
+    {
+        let mut ids = Box::new([0; PacketIdentifiers::SIZE]);
+        for i in 0..PacketIdentifiers::SIZE {
+            ids[i] = match seq.next_element()? {
+                Some(val) => val,
+                None => return Err(serde::de::Error::invalid_length(i, &self)),
+            };
+        }
+        Ok(IdentifiersInUse(ids))
+    }
+}
+
+impl Serialize for IdentifiersInUse {
+    #[inline]
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let mut seq = serializer.serialize_tuple(PacketIdentifiers::SIZE)?;
+        for e in self.0.iter() {
+            seq.serialize_element(e)?;
+        }
+        seq.end()
+    }
+}
+
+impl<'de> Deserialize<'de> for IdentifiersInUse {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        deserializer.deserialize_tuple(PacketIdentifiers::SIZE, IdentifiersInUseVisitor)
+    }
+}
+
+#[derive(Clone, PartialEq, Serialize, Deserialize)]
+pub(crate) struct PacketIdentifiers {
+    in_use: IdentifiersInUse,
+    previous: proto::PacketIdentifier,
+}
+
+impl PacketIdentifiers {
+    /// Size of a bitset for every packet identifier
+    ///
+    /// Packet identifiers are u16's, so the number of usize's required
+    /// = number of u16's / number of bits in a usize
+    /// = pow(2, number of bits in a u16) / number of bits in a usize
+    /// = pow(2, 16) / (`size_of::<usize>()` * 8)
+    ///
+    /// We use a bitshift instead of `usize::pow` because the latter is not a const fn
+    pub(crate) const SIZE: usize = (1 << 16) / (mem::size_of::<usize>() * 8);
+
+    #[cfg(any(test, feature = "proptest"))]
+    pub(crate) fn new(in_use: IdentifiersInUse, previous: proto::PacketIdentifier) -> Self {
+        Self { in_use, previous }
+    }
+
+    pub(crate) fn reserve(&mut self) -> Result<proto::PacketIdentifier, Error> {
+        let start = self.previous;
+        let mut current = start;
+
+        current += 1;
+
+        let (block, mask) = self.entry(current);
+        if (*block & mask) != 0 {
+            return Err(Error::PacketIdentifiersExhausted);
+        }
+
+        *block |= mask;
+        self.previous = current;
+        Ok(current)
+    }
+
+    pub(crate) fn discard(&mut self, packet_identifier: proto::PacketIdentifier) {
+        let (block, mask) = self.entry(packet_identifier);
+        *block &= !mask;
+    }
+
+    pub(crate) fn entry(
+        &mut self,
+        packet_identifier: proto::PacketIdentifier,
+    ) -> (&mut usize, usize) {
+        let packet_identifier = usize::from(packet_identifier.get());
+        let (block, offset) = (
+            packet_identifier / (mem::size_of::<usize>() * 8),
+            packet_identifier % (mem::size_of::<usize>() * 8),
+        );
+        (&mut self.in_use.0[block], 1 << offset)
+    }
+}
+
+impl fmt::Debug for PacketIdentifiers {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("PacketIdentifiers")
+            .field("previous", &self.previous)
+            .finish()
+    }
+}
+
+impl Default for PacketIdentifiers {
+    fn default() -> Self {
+        PacketIdentifiers {
+            in_use: IdentifiersInUse(Box::new([0; PacketIdentifiers::SIZE])),
+            previous: proto::PacketIdentifier::max_value(),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::PacketIdentifiers;
+
+    #[test]
+    fn packet_identifiers() {
+        #[cfg(target_pointer_width = "32")]
+        assert_eq!(PacketIdentifiers::SIZE, 2048);
+        #[cfg(target_pointer_width = "64")]
+        assert_eq!(PacketIdentifiers::SIZE, 1024);
+
+        let mut packet_identifiers = PacketIdentifiers::default();
+        assert_eq!(
+            packet_identifiers.in_use.0[..],
+            Box::new([0; PacketIdentifiers::SIZE])[..]
+        );
+
+        assert_eq!(packet_identifiers.reserve().unwrap().get(), 1);
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        expected[0] = 1 << 1;
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+
+        assert_eq!(packet_identifiers.reserve().unwrap().get(), 2);
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        expected[0] = (1 << 1) | (1 << 2);
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+
+        assert_eq!(packet_identifiers.reserve().unwrap().get(), 3);
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        expected[0] = (1 << 1) | (1 << 2) | (1 << 3);
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+
+        packet_identifiers.discard(crate::proto::PacketIdentifier::new(2).unwrap());
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        expected[0] = (1 << 1) | (1 << 3);
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+
+        assert_eq!(packet_identifiers.reserve().unwrap().get(), 4);
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        expected[0] = (1 << 1) | (1 << 3) | (1 << 4);
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+
+        packet_identifiers.discard(crate::proto::PacketIdentifier::new(1).unwrap());
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        expected[0] = (1 << 3) | (1 << 4);
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+
+        packet_identifiers.discard(crate::proto::PacketIdentifier::new(3).unwrap());
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        expected[0] = 1 << 4;
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+
+        packet_identifiers.discard(crate::proto::PacketIdentifier::new(4).unwrap());
+        assert_eq!(
+            packet_identifiers.in_use.0[..],
+            Box::new([0; PacketIdentifiers::SIZE])[..]
+        );
+
+        assert_eq!(packet_identifiers.reserve().unwrap().get(), 5);
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        expected[0] = 1 << 5;
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+
+        let goes_in_next_block = std::mem::size_of::<usize>() * 8;
+        #[allow(clippy::cast_possible_truncation)]
+        for i in 6..=goes_in_next_block {
+            assert_eq!(packet_identifiers.reserve().unwrap().get(), i as u16);
+        }
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        #[allow(clippy::identity_op)]
+        {
+            expected[0] = usize::max_value() - (1 << 0) - (1 << 1) - (1 << 2) - (1 << 3) - (1 << 4);
+            expected[1] |= 1 << 0;
+        }
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+
+        #[allow(clippy::cast_possible_truncation, clippy::range_minus_one)]
+        for i in 5..=(goes_in_next_block - 1) {
+            packet_identifiers.discard(crate::proto::PacketIdentifier::new(i as u16).unwrap());
+        }
+        let mut expected = Box::new([0; PacketIdentifiers::SIZE]);
+        #[allow(clippy::identity_op)]
+        {
+            expected[1] |= 1 << 0;
+        }
+        assert_eq!(packet_identifiers.in_use.0[..], expected[..]);
+    }
+}
