+++ mqtt/mqtt-broker/src/session/mod.rs
@@ -1,0 +1,424 @@
+mod connected;
+mod disconnecting;
+pub(crate) mod identifiers;
+mod offline;
+mod state;
+
+pub use connected::ConnectedSession;
+use disconnecting::DisconnectingSession;
+use offline::OfflineSession;
+pub use state::SessionState;
+
+use std::collections::HashMap;
+
+use mqtt3::proto;
+use mqtt_broker_core::{auth::AuthId, ClientInfo};
+
+use crate::subscription::Subscription;
+use crate::{ClientEvent, ClientId, ConnReq, ConnectionHandle, Error};
+
+#[derive(Debug)]
+pub enum Session {
+    Transient(ConnectedSession),
+    Persistent(ConnectedSession),
+    Disconnecting(DisconnectingSession),
+    Offline(OfflineSession),
+}
+
+impl Session {
+    pub fn new_transient(auth_id: AuthId, connreq: ConnReq, state: SessionState) -> Self {
+        let (peer_addr, connect, handle) = connreq.into_parts();
+        let client_info = ClientInfo::new(peer_addr, auth_id);
+        let connected = ConnectedSession::new(state, client_info, connect.will, handle);
+        Self::Transient(connected)
+    }
+
+    pub fn new_persistent(auth_id: AuthId, connreq: ConnReq, state: SessionState) -> Self {
+        let (peer_addr, connect, handle) = connreq.into_parts();
+        let client_info = ClientInfo::new(peer_addr, auth_id);
+        let connected = ConnectedSession::new(state, client_info, connect.will, handle);
+        Self::Persistent(connected)
+    }
+
+    pub fn new_offline(state: SessionState) -> Self {
+        let offline = OfflineSession::new(state);
+        Self::Offline(offline)
+    }
+
+    pub fn new_disconnecting(
+        client_id: ClientId,
+        client_info: ClientInfo,
+        will: Option<proto::Publication>,
+        handle: ConnectionHandle,
+    ) -> Self {
+        let disconnecting = DisconnectingSession::new(client_id, client_info, will, handle);
+        Self::Disconnecting(disconnecting)
+    }
+
+    pub fn client_id(&self) -> &ClientId {
+        match self {
+            Self::Transient(connected) => connected.client_id(),
+            Self::Persistent(connected) => connected.client_id(),
+            Self::Offline(offline) => offline.client_id(),
+            Self::Disconnecting(disconnecting) => disconnecting.client_id(),
+        }
+    }
+
+    pub fn client_info(&self) -> Result<&ClientInfo, Error> {
+        match self {
+            Self::Transient(connected) => Ok(connected.client_info()),
+            Self::Persistent(connected) => Ok(connected.client_info()),
+            Self::Offline(_offline) => Err(Error::SessionOffline),
+            Self::Disconnecting(disconnecting) => Ok(disconnecting.client_info()),
+        }
+    }
+
+    pub fn into_will(self) -> Option<proto::Publication> {
+        match self {
+            Self::Transient(connected) => connected.into_will(),
+            Self::Persistent(connected) => connected.into_will(),
+            Self::Offline(_offline) => None,
+            Self::Disconnecting(disconnecting) => disconnecting.into_will(),
+        }
+    }
+
+    pub fn subscriptions(&self) -> Option<&HashMap<String, Subscription>> {
+        match self {
+            Self::Transient(connected) => Some(connected.subscriptions()),
+            Self::Persistent(connected) => Some(connected.subscriptions()),
+            Self::Offline(offline) => Some(offline.subscriptions()),
+            Self::Disconnecting(_) => None,
+        }
+    }
+
+    pub fn handle_publish(
+        &mut self,
+        publish: proto::Publish,
+    ) -> Result<(Option<proto::Publication>, Option<ClientEvent>), Error> {
+        match self {
+            Self::Transient(connected) => connected.handle_publish(publish),
+            Self::Persistent(connected) => connected.handle_publish(publish),
+            Self::Offline(_offline) => Err(Error::SessionOffline),
+            Self::Disconnecting(_) => Err(Error::SessionOffline),
+        }
+    }
+
+    pub fn handle_puback(&mut self, puback: &proto::PubAck) -> Result<Option<ClientEvent>, Error> {
+        match self {
+            Self::Transient(connected) => connected.handle_puback(puback),
+            Self::Persistent(connected) => connected.handle_puback(puback),
+            Self::Offline(_offline) => Err(Error::SessionOffline),
+            Self::Disconnecting(_) => Err(Error::SessionOffline),
+        }
+    }
+
+    pub fn handle_puback0(
+        &mut self,
+        id: proto::PacketIdentifier,
+    ) -> Result<Option<ClientEvent>, Error> {
+        match self {
+            Self::Transient(connected) => connected.handle_puback0(id),
+            Self::Persistent(connected) => connected.handle_puback0(id),
+            Self::Offline(_offline) => Err(Error::SessionOffline),
+            Self::Disconnecting(_) => Err(Error::SessionOffline),
+        }
+    }
+
+    pub fn handle_pubrec(&mut self, pubrec: &proto::PubRec) -> Result<Option<ClientEvent>, Error> {
+        match self {
+            Self::Transient(connected) => connected.handle_pubrec(pubrec),
+            Self::Persistent(connected) => connected.handle_pubrec(pubrec),
+            Self::Offline(_offline) => Err(Error::SessionOffline),
+            Self::Disconnecting(_) => Err(Error::SessionOffline),
+        }
+    }
+
+    pub fn handle_pubrel(
+        &mut self,
+        pubrel: &proto::PubRel,
+    ) -> Result<Option<proto::Publication>, Error> {
+        match self {
+            Self::Transient(connected) => connected.handle_pubrel(pubrel),
+            Self::Persistent(connected) => connected.handle_pubrel(pubrel),
+            Self::Offline(_offline) => Err(Error::SessionOffline),
+            Self::Disconnecting(_) => Err(Error::SessionOffline),
+        }
+    }
+
+    pub fn handle_pubcomp(
+        &mut self,
+        pubcomp: &proto::PubComp,
+    ) -> Result<Option<ClientEvent>, Error> {
+        match self {
+            Self::Transient(connected) => connected.handle_pubcomp(pubcomp),
+            Self::Persistent(connected) => connected.handle_pubcomp(pubcomp),
+            Self::Offline(_offline) => Err(Error::SessionOffline),
+            Self::Disconnecting(_) => Err(Error::SessionOffline),
+        }
+    }
+
+    pub fn publish_to(
+        &mut self,
+        publication: &proto::Publication,
+    ) -> Result<Option<ClientEvent>, Error> {
+        match self {
+            Self::Transient(connected) => connected.publish_to(publication.to_owned()),
+            Self::Persistent(connected) => connected.publish_to(publication.to_owned()),
+            Self::Offline(offline) => offline.publish_to(publication.to_owned()),
+            Self::Disconnecting(_) => Err(Error::SessionOffline),
+        }
+    }
+
+    pub fn subscribe_to(
+        &mut self,
+        subscribe_to: proto::SubscribeTo,
+    ) -> Result<(proto::SubAckQos, Option<Subscription>), Error> {
+        match self {
+            Self::Transient(connected) => connected.subscribe_to(subscribe_to),
+            Self::Persistent(connected) => connected.subscribe_to(subscribe_to),
+            Self::Offline(_) => Err(Error::SessionOffline),
+            Self::Disconnecting(_) => Err(Error::SessionOffline),
+        }
+    }
+
+    pub fn unsubscribe(
+        &mut self,
+        unsubscribe: &proto::Unsubscribe,
+    ) -> Result<proto::UnsubAck, Error> {
+        match self {
+            Self::Transient(connected) => connected.unsubscribe(unsubscribe),
+            Self::Persistent(connected) => connected.unsubscribe(unsubscribe),
+            Self::Offline(_) => Err(Error::SessionOffline),
+            Self::Disconnecting(_) => Err(Error::SessionOffline),
+        }
+    }
+
+    pub fn send(&mut self, event: ClientEvent) -> Result<(), Error> {
+        match self {
+            Self::Transient(ref mut connected) => connected.send(event),
+            Self::Persistent(ref mut connected) => connected.send(event),
+            Self::Disconnecting(ref mut disconnecting) => disconnecting.send(event),
+            _ => Err(Error::SessionOffline),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::time::Duration;
+
+    use matches::assert_matches;
+    use tokio::sync::mpsc;
+    use uuid::Uuid;
+
+    use mqtt3::{proto, PROTOCOL_LEVEL, PROTOCOL_NAME};
+    use mqtt_broker_core::auth::AuthId;
+
+    use super::{Session, SessionState};
+    use crate::{
+        configuration::QueueFullAction, tests::peer_addr, Auth, ClientId, ConnReq,
+        ConnectionHandle, Error, SessionConfig,
+    };
+
+    fn connection_handle() -> ConnectionHandle {
+        let id = Uuid::new_v4();
+        let (tx1, _rx1) = mpsc::unbounded_channel();
+        ConnectionHandle::new(id, tx1)
+    }
+
+    fn transient_connect(id: String) -> proto::Connect {
+        proto::Connect {
+            username: None,
+            password: None,
+            will: None,
+            client_id: proto::ClientId::IdWithCleanSession(id),
+            keep_alive: Duration::default(),
+            protocol_name: PROTOCOL_NAME.to_string(),
+            protocol_level: PROTOCOL_LEVEL,
+        }
+    }
+
+    fn default_config() -> SessionConfig {
+        SessionConfig::new(
+            Duration::default(),
+            0,
+            16,
+            1000,
+            0,
+            QueueFullAction::DropNew,
+        )
+    }
+
+    #[test]
+    fn test_subscribe_to() {
+        let id = "id1".to_string();
+        let client_id = ClientId::from(id.clone());
+        let connect1 = transient_connect(id);
+        let handle1 = connection_handle();
+        let req1 = ConnReq::new(
+            client_id.clone(),
+            peer_addr(),
+            connect1,
+            Auth::Unknown,
+            handle1,
+        );
+        let auth_id = "auth-id1".into();
+        let state = SessionState::new(client_id, default_config());
+        let mut session = Session::new_transient(auth_id, req1, state);
+        let subscribe_to = proto::SubscribeTo {
+            topic_filter: "topic/new".to_string(),
+            qos: proto::QoS::AtMostOnce,
+        };
+
+        let (ack, subscription) = session.subscribe_to(subscribe_to).unwrap();
+
+        assert_eq!(ack, proto::SubAckQos::Success(proto::QoS::AtMostOnce));
+        assert_matches!(subscription, Some(_));
+        match &session {
+            Session::Transient(connected) => {
+                assert_eq!(1, connected.subscriptions().len());
+                assert_eq!(
+                    proto::QoS::AtMostOnce,
+                    *connected.subscriptions()["topic/new"].max_qos()
+                );
+            }
+            _ => panic!("not transient"),
+        }
+
+        let subscribe_to = proto::SubscribeTo {
+            topic_filter: "topic/new".to_string(),
+            qos: proto::QoS::AtLeastOnce,
+        };
+
+        let (ack, subscription) = session.subscribe_to(subscribe_to).unwrap();
+
+        assert_eq!(ack, proto::SubAckQos::Success(proto::QoS::AtLeastOnce));
+        assert_matches!(subscription, Some(_));
+        match &session {
+            Session::Transient(connected) => {
+                assert_eq!(1, connected.subscriptions().len());
+                assert_eq!(
+                    proto::QoS::AtLeastOnce,
+                    *connected.subscriptions()["topic/new"].max_qos()
+                );
+            }
+            _ => panic!("not transient"),
+        }
+    }
+
+    #[test]
+    fn test_subscribe_to_with_invalid_topic() {
+        let id = "id1".to_string();
+        let client_id = ClientId::from(id.clone());
+        let connect1 = transient_connect(id);
+        let handle1 = connection_handle();
+        let req1 = ConnReq::new(
+            client_id.clone(),
+            peer_addr(),
+            connect1,
+            Auth::Unknown,
+            handle1,
+        );
+        let auth_id = "auth-id1".into();
+        let state = SessionState::new(client_id, default_config());
+        let mut session = Session::new_transient(auth_id, req1, state);
+        let subscribe_to = proto::SubscribeTo {
+            topic_filter: "topic/#/#".to_string(),
+            qos: proto::QoS::AtMostOnce,
+        };
+
+        let (ack, subscription) = session.subscribe_to(subscribe_to).unwrap();
+
+        assert_eq!(ack, proto::SubAckQos::Failure);
+        assert_eq!(subscription, None);
+    }
+
+    #[test]
+    fn test_unsubscribe() {
+        let id = "id1".to_string();
+        let client_id = ClientId::from(id.clone());
+        let connect1 = transient_connect(id);
+        let handle1 = connection_handle();
+        let req1 = ConnReq::new(
+            client_id.clone(),
+            peer_addr(),
+            connect1,
+            Auth::Unknown,
+            handle1,
+        );
+        let auth_id = AuthId::Anonymous;
+        let state = SessionState::new(client_id, default_config());
+        let mut session = Session::new_transient(auth_id, req1, state);
+
+        let subscribe_to = proto::SubscribeTo {
+            topic_filter: "topic/new".to_string(),
+            qos: proto::QoS::AtMostOnce,
+        };
+
+        let (ack, subscription) = session.subscribe_to(subscribe_to).unwrap();
+        assert_eq!(ack, proto::SubAckQos::Success(proto::QoS::AtMostOnce));
+        assert_matches!(subscription, Some(_));
+
+        let unsubscribe = proto::Unsubscribe {
+            packet_identifier: proto::PacketIdentifier::new(1).unwrap(),
+            unsubscribe_from: vec!["topic/different".to_string()],
+        };
+        session.unsubscribe(&unsubscribe).unwrap();
+
+        match &session {
+            Session::Transient(connected) => {
+                assert_eq!(1, connected.subscriptions().len());
+                assert_eq!(
+                    proto::QoS::AtMostOnce,
+                    *connected.subscriptions()["topic/new"].max_qos()
+                );
+            }
+            _ => panic!("not transient"),
+        }
+
+        let unsubscribe = proto::Unsubscribe {
+            packet_identifier: proto::PacketIdentifier::new(24).unwrap(),
+            unsubscribe_from: vec!["topic/new".to_string()],
+        };
+        let unsuback = session.unsubscribe(&unsubscribe).unwrap();
+        assert_eq!(
+            proto::PacketIdentifier::new(24).unwrap(),
+            unsuback.packet_identifier
+        );
+
+        match &session {
+            Session::Transient(connected) => {
+                assert_eq!(0, connected.subscriptions().len());
+            }
+            _ => panic!("not transient"),
+        }
+    }
+
+    #[test]
+    fn test_offline_subscribe_to() {
+        let id = "id1".to_string();
+        let client_id = ClientId::from(id);
+        let mut session = Session::new_offline(SessionState::new(client_id, default_config()));
+
+        let subscribe_to = proto::SubscribeTo {
+            topic_filter: "topic/new".to_string(),
+            qos: proto::QoS::AtMostOnce,
+        };
+        let result = session.subscribe_to(subscribe_to);
+        assert_matches!(result, Err(Error::SessionOffline));
+    }
+
+    #[test]
+    fn test_offline_unsubscribe() {
+        let id = "id1".to_string();
+        let client_id = ClientId::from(id);
+        let mut session = Session::new_offline(SessionState::new(client_id, default_config()));
+
+        let unsubscribe = proto::Unsubscribe {
+            packet_identifier: proto::PacketIdentifier::new(24).unwrap(),
+            unsubscribe_from: vec!["topic/new".to_string()],
+        };
+        let result = session.unsubscribe(&unsubscribe);
+        assert_matches!(result, Err(Error::SessionOffline));
+    }
+}
